Renamed the projectTabs folder in the modules directory and updated all import paths where the folder is used.
Renamed the projectTabs folder in the modules directory and updated all import paths where the folder is used.
