+++ modules/project-tabs/controller.js
