+++ modules/project-tabs/init.js
